Backend de catalgo productos
Pues borre lo que ya no ocupabamos ahora lo que hice fue agregar para que se puedan ver los productos me confirmas si necesitas que desde el front se puedan borrar y editar
