+++ app/admin/layout.tsx
@@ -2,7 +2,7 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import React from "react";
-import { Home, Users, Building2, BarChart3, Settings } from "lucide-react";
+import { Home, Users, Building2, BarChart3, Settings, ShoppingCart, PackageSearch } from "lucide-react";
 
 export default function AdminLayout({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
@@ -10,9 +10,9 @@ export default function AdminLayout({ children }: { children: React.ReactNode })
   const links = [
     { href: "/admin", label: "Dashboard", icon: <Home size={18} /> },
     { href: "/admin/clientes", label: "Clientes", icon: <Users size={18} /> },
-    { href: "/admin/empresas", label: "Empresas", icon: <Building2 size={18} /> },
+    { href: "/admin/em", label: "Empresas", icon: <Building2 size={18} /> },
     { href: "/admin/reportes", label: "Reportes", icon: <BarChart3 size={18} /> },
-    { href: "/admin/configuracion", label: "Configuración", icon: <Settings size={18} /> },
+    { href: "/admin/productos", label: "Productos", icon: <PackageSearch size={18} /> },
   ];
 
   return (
